+++ src/test/groovy/edu/ie3/test/common/SampleJointGrid.groovy
@@ -77,7 +77,6 @@ class SampleJointGrid extends SystemParticipantTestData {
         elevationAngle,
         kG,
         kT,
-        false,
         sRated,
         cosPhiRated)