+++ src/test/groovy/edu/ie3/test/common/SystemParticipantTestData.groovy
@@ -25,7 +25,6 @@ import edu.ie3.datamodel.models.input.thermal.ThermalBusInput
 import edu.ie3.datamodel.models.input.thermal.ThermalStorageInput
 import edu.ie3.datamodel.models.profile.BdewStandardLoadProfile
 import edu.ie3.datamodel.models.profile.PowerProfileKey
-import edu.ie3.datamodel.models.profile.StandardLoadProfile
 import edu.ie3.util.TimeUtil
 import edu.ie3.util.quantities.interfaces.*
 import tech.units.indriya.ComparableQuantity
@@ -110,7 +109,6 @@ class SystemParticipantTestData {
   elevationAngle,
   kG,
   kT,
-  false,
   sRated,
   cosPhiRated
   )